Add files via upload
Rule-breaking palettes are celebrated in a new book about daring graphic design colour schemes

Colours are a fundamental part of graphic design, but even using them in a seemingly 'incorrect' way can produce striking, eye-catching results. And it's these colour schemes that rip up the rule book, which is the focus of a new book recently released by Counterprint.